Bump parent to pom-scijava 29.2.1
This increases the TrackMate version to 6.0.0, which changes the handling of priorities to follow standard SciJava priorities.

Therefore, we need to update the priority of LinearTrackDescriptor to Priority.LOW

While at it, let's change the test to use the actual runtime priorities. The adapted test fails with the previous priority=1d, but passes with priority=Priority.LOW.